clippy fix

+++ dkg-gadget/src/worker.rs
@@ -1079,7 +1079,7 @@ where
 
 		let mut offenders: Vec<AuthorityId> = Vec::new();
 		for bad_actor in bad_actors {
-			let bad_actor = bad_actor as usize;
+			let bad_actor = bad_actor;
 			if bad_actor > 0 && bad_actor <= authorities.len() {
 				if let Some(offender) = authorities.get(bad_actor - 1) {
 					offenders.push(offender.clone());
